requirements.txt
# this is the requirements file
requests

plotly

pandas

python-dotenv

pytest